image/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
boto3
pillow
redis
